Add a dry-run mode to deploy_service.rb.
When the environment variable FASTLY_DRY_RUN is set to anything other
than empty string, 'false' or 0, just print out the config instead of
uploading it to Fastly and activating it.

+++ lib/deploy_service.rb
@@ -14,6 +14,13 @@ class DeployService
     puts "Environment: #{environment}"
 
     vcl = RenderTemplate.render_template(configuration, environment, config, version)
+
+    dry_run = ENV.fetch("FASTLY_DRY_RUN", "").downcase
+    unless ["", "0", "false"].include?(dry_run)
+      puts vcl
+      exit
+    end
+
     delete_ui_objects(service.id, version.number)
     upload_vcl(version, vcl)
     diff_vcl(service, version)
